docs(readme): Update sa-webapp README files
- 顺带更新 Sentiment 结构的 Sentiments 的数据类型为 float64

+++ sa-webapp/webapp.go
@@ -21,7 +21,7 @@ type AnalyseSentimentData struct {
 
 type Sentiment struct {
 	Sentence   string  `json:"sentence"`
-	Sentiments float32 `json:"sentiments"`
+	Sentiments float64 `json:"sentiments"`
 }
 
 func sendAnalyseRequest(reqData *AnalyseSentimentData) ([]byte, error) {
